Move schema metadata yaml to infra module
