+++ shardingsphere-infra/shardingsphere-infra-common/src/main/java/org/apache/shardingsphere/infra/yaml/schema/pojo/YamlTableMetaData.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.governance.core.yaml.schema.pojo;
+package org.apache.shardingsphere.infra.yaml.schema.pojo;
 
 import lombok.Getter;
 import lombok.Setter;
@@ -24,11 +24,13 @@ import org.apache.shardingsphere.infra.yaml.config.pojo.YamlConfiguration;
 import java.util.Map;
 
 /**
- * ShardingSphere schema for YAML.
+ * Table meta data for YAML.
  */
 @Getter
 @Setter
-public final class YamlSchema implements YamlConfiguration {
+public final class YamlTableMetaData implements YamlConfiguration {
     
-    private Map<String, YamlTableMetaData> tables;
+    private Map<String, YamlColumnMetaData> columns;
+    
+    private Map<String, YamlIndexMetaData> indexes;
 }
