+++ shardingsphere-infra/shardingsphere-infra-common/src/test/java/org/apache/shardingsphere/infra/yaml/schema/swapper/SchemaYamlSwapperTest.java
@@ -15,12 +15,12 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.governance.core.yaml.schema.swapper;
+package org.apache.shardingsphere.infra.yaml.schema.swapper;
 
 import lombok.SneakyThrows;
-import org.apache.shardingsphere.governance.core.yaml.schema.pojo.YamlSchema;
 import org.apache.shardingsphere.infra.metadata.schema.ShardingSphereSchema;
 import org.apache.shardingsphere.infra.yaml.engine.YamlEngine;
+import org.apache.shardingsphere.infra.yaml.schema.pojo.YamlSchema;
 import org.junit.Test;
 
 import java.io.IOException;
@@ -36,9 +36,9 @@ import static org.junit.Assert.assertThat;
 
 public final class SchemaYamlSwapperTest {
     
-    private static final String YAML = "yaml/schema.yaml";
+    private static final String YAML = "yaml/schema/schema.yaml";
     
-    private static final String YAML_WITHOUT_TABLE = "yaml/schema-without-table.yaml";
+    private static final String YAML_WITHOUT_TABLE = "yaml/schema/schema-without-table.yaml";
     
     @Test
     public void assertSwapToYamlSchema() {
