+++ shardingsphere-infra/shardingsphere-infra-common/src/test/resources/yaml/schema/schema-without-table.yaml
